+++ source/include/sf2d/sf2d_draw.c
@@ -2,7 +2,7 @@
 #include "sf2d_private.h"
 #include <math.h>
 
-void sf2d_draw_line(int x0, int y0, int x1, int y1, u32 color)
+void sf2d_draw_line(float x0, float y0, float x1, float y1, u32 color)
 {
 	sf2d_vertex_pos_col *vertices = sf2d_pool_malloc(4 * sizeof(sf2d_vertex_pos_col));
 	if (!vertices) return;
@@ -47,7 +47,7 @@ void sf2d_draw_line(int x0, int y0, int x1, int y1, u32 color)
 	GPU_DrawArray(GPU_TRIANGLE_STRIP, 0, 4);
 }
 
-void sf2d_draw_rectangle(int x, int y, int w, int h, u32 color)
+void sf2d_draw_rectangle(float x, float y, float w, float h, u32 color)
 {
 	sf2d_vertex_pos_col *vertices = sf2d_pool_malloc(4 * sizeof(sf2d_vertex_pos_col));
 	if (!vertices) return;
@@ -92,13 +92,13 @@ void sf2d_draw_rectangle(int x, int y, int w, int h, u32 color)
 	GPU_DrawArray(GPU_TRIANGLE_STRIP, 0, 4);
 }
 
-void sf2d_draw_rectangle_rotate(int x, int y, int w, int h, u32 color, float rad)
+void sf2d_draw_rectangle_rotate(float x, float y, float w, float h, u32 color, float rad)
 {
 	sf2d_vertex_pos_col *vertices = sf2d_pool_malloc(4 * sizeof(sf2d_vertex_pos_col));
 	if (!vertices) return;
 
-	int w2 = w/2.0f;
-	int h2 = h/2.0f;
+	float w2 = w/2.0f;
+	float h2 = h/2.0f;
 
 	vertices[0].position = (sf2d_vector_3f){(float)-w2, (float)-h2, SF2D_DEFAULT_DEPTH};
 	vertices[1].position = (sf2d_vector_3f){(float) w2, (float)-h2, SF2D_DEFAULT_DEPTH};
@@ -150,7 +150,7 @@ void sf2d_draw_rectangle_rotate(int x, int y, int w, int h, u32 color, float rad
 	GPU_DrawArray(GPU_TRIANGLE_STRIP, 0, 4);
 }
 
-void sf2d_draw_fill_circle(int x, int y, int radius, u32 color)
+void sf2d_draw_fill_circle(float x, float y, int radius, u32 color)
 {
 	static const int num_segments = 100;
 	sf2d_vertex_pos_col *vertices = sf2d_pool_malloc((num_segments + 2) * sizeof(sf2d_vertex_pos_col));
